Move color helpers to separate files
- Extracted `colorToHex` and `hexToColor` from `calendar.go` into `color.go`.
- Extracted `TestColorToHex` and `TestHexToColor` from `calendar_test.go` into `color_test.go`.
